# feat(tsdb/(head|agent)): reuse pools across segments to reduce garbage during WL replay