@machine424 2025-02-12
Runs it with different wal files


```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
               │   v1.txt   │              v2.txt              │
               │   sec/op   │   sec/op    vs base              │
LoadRealWLs-26   125.1 ± 1%   123.2 ± 2%  -1.53% (p=0.041 n=6)

               │    v1.txt    │               v2.txt               │
               │     B/op     │     B/op      vs base              │
LoadRealWLs-26   31.98Gi ± 0%   30.05Gi ± 1%  -6.03% (p=0.002 n=6)

               │   v1.txt    │            v2.txt            │
               │  allocs/op  │  allocs/op   vs base         │
LoadRealWLs-26   482.3M ± 0%   482.2M ± 0%  ~ (p=0.093 n=6)
``` 

Ran more benchmarks, no remarkable changes as usually only 1 segment is involved:

<details>

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Intel(R) Xeon(R) Platinum 8175M CPU @ 2.50GHz
                                                                                                                                                             │   v11.txt    │              v22.txt               │
                                                                                                                                                             │    sec/op    │    sec/op     vs base              │
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26          1.418 ± 16%    1.497 ±  7%       ~ (p=0.485 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         1.436 ±  5%    1.455 ±  2%       ~ (p=0.132 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         1.447 ± 10%    1.502 ±  8%       ~ (p=0.394 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        1.506 ±  5%    1.561 ±  6%       ~ (p=0.180 n=6)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         660.4m ±  4%   654.7m ±  7%       ~ (p=0.485 n=6)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         711.8m ±  2%   717.8m ±  3%       ~ (p=0.240 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         530.9m ± 11%   546.1m ±  2%       ~ (p=0.394 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         551.7m ±  3%   538.9m ±  5%       ~ (p=0.699 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         569.8m ±  7%   558.9m ± 23%       ~ (p=0.818 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        579.7m ± 10%   582.3m ± 16%       ~ (p=0.937 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26      2.866 ±  8%    2.901 ±  6%       ~ (p=0.699 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26      2.935 ±  9%    2.917 ±  3%       ~ (p=0.937 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26      2.890 ±  7%    2.994 ±  9%       ~ (p=0.180 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26     3.316 ±  8%    3.506 ±  6%       ~ (p=0.132 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26         2.864 ±  4%    2.840 ±  7%       ~ (p=0.485 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26         2.844 ±  6%    2.925 ±  5%       ~ (p=0.310 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26         2.920 ±  5%    2.916 ±  6%       ~ (p=0.699 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26        2.925 ±  5%    2.924 ±  2%       ~ (p=0.699 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26          1.420 ±  7%    1.433 ±  7%       ~ (p=0.394 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26          1.395 ±  4%    1.395 ±  7%       ~ (p=0.937 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26          1.404 ± 11%    1.412 ±  5%       ~ (p=0.589 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26         1.409 ±  7%    1.479 ±  6%       ~ (p=0.180 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26     11.44 ±  3%    11.06 ±  4%       ~ (p=0.093 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26     11.39 ±  1%    11.38 ±  2%       ~ (p=1.000 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26     11.39 ±  3%    11.35 ±  3%       ~ (p=0.937 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26    11.78 ±  2%    11.68 ±  2%       ~ (p=0.310 n=6)
geomean                                                                                                                                                         2.011          2.031        +1.02%

                                                                                                                                                             │   v11.txt    │              v22.txt               │
                                                                                                                                                             │     B/op     │     B/op      vs base              │
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         69.49Mi ± 0%   69.47Mi ± 0%       ~ (p=0.394 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        71.29Mi ± 0%   71.29Mi ± 0%       ~ (p=0.937 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        72.97Mi ± 0%   72.95Mi ± 0%       ~ (p=0.699 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26       88.35Mi ± 0%   88.36Mi ± 0%       ~ (p=0.485 n=6)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         243.0Mi ± 1%   242.2Mi ± 1%       ~ (p=0.485 n=6)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         267.0Mi ± 1%   270.0Mi ± 2%       ~ (p=0.132 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         71.29Mi ± 0%   71.49Mi ± 0%       ~ (p=0.132 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         73.22Mi ± 0%   73.26Mi ± 0%       ~ (p=0.818 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         74.50Mi ± 1%   75.12Mi ± 0%       ~ (p=0.065 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        84.50Mi ± 2%   85.08Mi ± 2%       ~ (p=0.132 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26     246.9Mi ± 1%   244.8Mi ± 1%  -0.82% (p=0.009 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26     260.4Mi ± 2%   256.7Mi ± 1%  -1.41% (p=0.026 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26     273.5Mi ± 1%   270.5Mi ± 1%  -1.11% (p=0.002 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26    360.6Mi ± 1%   358.9Mi ± 1%       ~ (p=0.240 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26        541.6Mi ± 0%   540.8Mi ± 0%  -0.14% (p=0.009 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26        542.8Mi ± 0%   542.4Mi ± 0%       ~ (p=0.394 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26        545.3Mi ± 0%   544.2Mi ± 0%  -0.20% (p=0.002 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26       554.1Mi ± 0%   552.8Mi ± 0%  -0.24% (p=0.015 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26         469.9Mi ± 0%   468.8Mi ± 0%       ~ (p=0.132 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26         471.7Mi ± 0%   470.6Mi ± 0%  -0.22% (p=0.041 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26         473.4Mi ± 0%   472.3Mi ± 0%       ~ (p=0.180 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26        482.7Mi ± 1%   481.1Mi ± 0%       ~ (p=0.065 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26    4.134Gi ± 0%   4.131Gi ± 0%  -0.06% (p=0.026 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26    4.147Gi ± 0%   4.141Gi ± 0%  -0.14% (p=0.009 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26    4.158Gi ± 0%   4.155Gi ± 0%  -0.08% (p=0.009 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26   4.246Gi ± 0%   4.239Gi ± 0%  -0.14% (p=0.002 n=6)
geomean                                                                                                                                                        339.8Mi        339.3Mi       -0.13%

                                                                                                                                                             │   v11.txt   │              v22.txt              │
                                                                                                                                                             │  allocs/op  │  allocs/op   vs base              │
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         687.6k ± 0%   687.6k ± 0%       ~ (p=0.589 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=36,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        796.0k ± 0%   796.0k ± 0%       ~ (p=0.394 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=72,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        904.3k ± 0%   904.3k ± 0%       ~ (p=0.485 n=6)
LoadWLs/batches=10,seriesPerBatch=100,samplesPerSeries=7200,exemplarsPerSeries=360,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26       1.771M ± 0%   1.771M ± 0%  +0.01% (p=0.041 n=6)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         2.064M ± 0%   2.064M ± 0%       ~ (p=0.240 n=6)
LoadWLs/batches=10,seriesPerBatch=10000,samplesPerSeries=50,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         2.664M ± 0%   2.664M ± 0%       ~ (p=0.065 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         644.5k ± 0%   644.6k ± 0%       ~ (p=0.240 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         704.7k ± 0%   704.6k ± 0%       ~ (p=0.818 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26         794.7k ± 0%   794.7k ± 0%       ~ (p=0.937 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26        1.365M ± 0%   1.365M ± 0%       ~ (p=0.699 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26     3.124M ± 0%   3.124M ± 0%       ~ (p=0.937 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26     3.736M ± 0%   3.727M ± 0%       ~ (p=0.180 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26     4.635M ± 0%   4.628M ± 0%  -0.15% (p=0.041 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.000,oooSamplesPct=0.000,oooCapMax=0-26    10.33M ± 0%   10.33M ± 0%       ~ (p=0.589 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26        2.904M ± 0%   2.903M ± 0%       ~ (p=0.937 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26        2.964M ± 0%   2.963M ± 0%       ~ (p=0.818 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26        3.054M ± 0%   3.054M ± 0%       ~ (p=0.485 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.500,oooSamplesPct=0.500,oooCapMax=32-26       3.623M ± 0%   3.623M ± 0%       ~ (p=0.937 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26         2.701M ± 0%   2.702M ± 0%       ~ (p=0.818 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26         2.762M ± 0%   2.761M ± 0%       ~ (p=0.180 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26         2.851M ± 0%   2.852M ± 0%       ~ (p=0.699 n=6)
LoadWLs/batches=10,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=0,oooSeriesPct=0.100,oooSamplesPct=0.100,oooCapMax=0-26        3.423M ± 0%   3.422M ± 0%       ~ (p=0.485 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=0,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26    23.90M ± 0%   23.91M ± 0%       ~ (p=0.180 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=2,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26    24.50M ± 0%   24.50M ± 0%       ~ (p=0.818 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=5,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26    25.40M ± 0%   25.40M ± 0%       ~ (p=0.589 n=6)
LoadWLs/batches=100,seriesPerBatch=1000,samplesPerSeries=480,exemplarsPerSeries=24,mmappedChunkT=3800,oooSeriesPct=0.200,oooSamplesPct=0.300,oooCapMax=32-26   31.11M ± 0%   31.10M ± 0%       ~ (p=0.093 n=6)
geomean                                                                                                                                                        3.060M        3.060M       -0.01%
```

</details>